test: wire all Swift test entry points
