+++ Sources/AppBuild.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+enum AppBuild {
+    static var isDevBundle: Bool {
+        isDevBundleName(Bundle.main.object(forInfoDictionaryKey: "CFBundleName") as? String)
+    }
+
+    static func isDevBundleName(_ bundleName: String?) -> Bool {
+        bundleName == "Quill Dev"
+    }
+}
