+++ Sources/AppState.swift
@@ -21,16 +21,6 @@ struct PrecomputedMacro {
     let normalizedCommand: String
 }
 
-enum AppBuild {
-    static var isDevBundle: Bool {
-        isDevBundleName(Bundle.main.object(forInfoDictionaryKey: "CFBundleName") as? String)
-    }
-
-    static func isDevBundleName(_ bundleName: String?) -> Bool {
-        bundleName == "Quill Dev"
-    }
-}
-
 private struct PreservedPasteboardEntry {
     let type: NSPasteboard.PasteboardType
     let value: Value
